RN-950 Fix cypress run error by removing @tupaia/e2e dependency on @tupaia/utils
The only caller is the e2e package. This somehow fixes issue RN-950
