+++ packages/e2e/cypress/support/utils.js
@@ -42,3 +42,12 @@ export const serializeReactToHTML = jQueryEl => {
 
   return beautifyHtml(sanitizedHtml, options);
 };
+
+export const requireCyEnv = variable => {
+  const unprefixedVariable = variable.replace(/^CYPRESS_/, '');
+  const value = Cypress.env(unprefixedVariable);
+  if (value === undefined) {
+    throw new Error(`Could not load Cypress env variable 'CYPRESS_${unprefixedVariable}'`);
+  }
+  return value;
+};
